DFS Implementation
This project models a 3D doughnut in which a fly is supposed to travel over.  A method of DFS is used to search the doughnut wherein only certain elements of the representational matrix are traversed.  A sample input file is given which shows the individual elements of the matrix or "modeled doughnut".